renames stats-model
